Refactor and test config.
