+++ src/main/java/uk/gov/ons/bulk/scheduler/component/PubSubOutboundComponent.java
@@ -1,0 +1,31 @@
+package uk.gov.ons.bulk.scheduler.component;
+
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.integration.annotation.MessagingGateway;
+import org.springframework.integration.annotation.ServiceActivator;
+import org.springframework.messaging.MessageHandler;
+import org.springframework.stereotype.Component;
+
+import com.google.cloud.spring.pubsub.core.PubSubTemplate;
+import com.google.cloud.spring.pubsub.integration.outbound.PubSubMessageHandler;
+
+@Component
+public class PubSubOutboundComponent {
+
+	@Value("${aims.pubsub.export-topic}")
+	private String pubsubExportTopic;
+
+	@Bean
+	@ServiceActivator(inputChannel = "pubsubOutputChannel")
+	public MessageHandler messageSender(PubSubTemplate pubsubTemplate) {
+		return new PubSubMessageHandler(pubsubTemplate, pubsubExportTopic);
+	}
+
+	@MessagingGateway(defaultRequestChannel = "pubsubOutputChannel")
+	public interface PubsubOutboundGateway {
+
+		void sendToPubsub(String text);
+	}
+}
+
