+++ src/test/resources/application-test.yml
@@ -1,6 +1,8 @@
 spring:
   autoconfigure:
-    exclude: org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration
+    exclude:
+      - org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration
+      - com.google.cloud.spring.autoconfigure.bigquery.GcpBigQueryAutoConfiguration
   cloud:
     gcp:
       project-id: ${embedded.google.pubsub.project-id}
@@ -10,3 +12,7 @@ spring:
         emulator-host: ${embedded.google.pubsub.host}:${embedded.google.pubsub.port}
       sql:
         enabled: false
+aims:
+  pubsub:
+    subscription: processing-finished-subscription-test
+    export-topic: bulk-scheduler-test
